feat(db): split misc.ts — PR2 per-source small domains + themes (Task 3.1)
Extract autoTraceroute (13 methods, 6 sync twins), timeSync (4), distanceDeleteLog (2),
mapPreferences (2), and themes (5) from MiscRepository into dedicated repositories.

- New repos: AutoTracerouteRepository, TimeSyncRepository, DistanceDeleteLogRepository,
  MapPreferencesRepository, ThemesRepository (all in src/db/repositories/)
- DatabaseService: 5 new fields + throwing getters + instantiations; repointed all
  facade delegations (autoTraceroute, timeSync, mapPreferences) off this.misc/miscRepo
- External callers: themeRoutes.ts → .themes; server.ts distanceDeleteLog → .distanceDeleteLog;
  autoDeleteByDistanceService.ts → .distanceDeleteLog
- Tests: autoTraceroute.test.ts (ported), timeSync.test.ts, distanceDeleteLog.test.ts,
  mapPreferences.test.ts (schema rename), themes.test.ts; perSource.test.ts repointed
  to 3 new repos; themeRoutes.test.ts mockDb.misc → mockDb.themes
- eslint-baseline.json updated to reflect verbatim-moved `any` count

Verified: tsc clean, lint:ci OK, build green, 8456 tests pass (0 failures).
