+++ src/db/repositories/misc.ts
@@ -1,424 +1,21 @@
 /**
  * Misc Repository
  *
- * Handles solar estimates and auto-traceroute nodes database operations.
+ * Handles packet log and key repair database operations.
  * Supports SQLite, PostgreSQL, and MySQL through Drizzle ORM.
  */
-import { eq, desc, asc, and, or, inArray, notInArray, sql, isNull } from 'drizzle-orm';
+import { eq, asc, and, or, inArray, sql, isNull } from 'drizzle-orm';
 import { BaseRepository, DrizzleDatabase } from './base.js';
-import { DatabaseType, DbCustomTheme, DbPacketLog, DbPacketCountByNode, DbPacketCountByPortnum, DbDistinctRelayNode } from '../types.js';
+import { DatabaseType, DbPacketLog, DbPacketCountByNode, DbPacketCountByPortnum, DbDistinctRelayNode } from '../types.js';
 import { logger } from '../../utils/logger.js';
 import { getPortNumName } from '../../server/constants/meshtastic.js';
 
-export interface AutoTracerouteNode {
-  id?: number;
-  nodeNum: number;
-  enabled?: boolean;
-  createdAt: number;
-}
-
 /**
- * Repository for miscellaneous operations (solar estimates, auto-traceroute nodes, news)
+ * Repository for miscellaneous operations (packet log, key repair)
  */
 export class MiscRepository extends BaseRepository {
   constructor(db: DrizzleDatabase, dbType: DatabaseType) {
     super(db, dbType);
-  }
-
-  // ============ AUTO-TRACEROUTE NODES ============
-
-  /**
-   * Get all auto-traceroute nodes.
-   * When sourceId is provided, return only rows scoped to that source OR
-   * legacy unscoped rows (sourceId IS NULL). When omitted, return everything.
-   */
-  async getAutoTracerouteNodes(sourceId?: string): Promise<number[]> {
-    const { autoTracerouteNodes } = this.tables;
-    const query = this.db
-      .select({ nodeNum: autoTracerouteNodes.nodeNum })
-      .from(autoTracerouteNodes);
-    const results = sourceId
-      ? await query
-          .where(eq(autoTracerouteNodes.sourceId, sourceId))
-          .orderBy(asc(autoTracerouteNodes.createdAt))
-      : await query.orderBy(asc(autoTracerouteNodes.createdAt));
-    return results.map((r: any) => Number(r.nodeNum));
-  }
-
-  /**
-   * Set auto-traceroute nodes (replaces all existing entries for the given
-   * source, or globally when sourceId is omitted).
-   */
-  async setAutoTracerouteNodes(nodeNums: number[], sourceId?: string): Promise<void> {
-    const now = this.now();
-    const { autoTracerouteNodes } = this.tables;
-
-    // Delete existing entries scoped to this source (or all when unscoped).
-    if (sourceId) {
-      await this.db
-        .delete(autoTracerouteNodes)
-        .where(eq(autoTracerouteNodes.sourceId, sourceId));
-    } else {
-      await this.db.delete(autoTracerouteNodes);
-    }
-    // Insert new entries
-    for (const nodeNum of nodeNums) {
-      await this.db
-        .insert(autoTracerouteNodes)
-        .values({ nodeNum, createdAt: now, sourceId: sourceId ?? null });
-    }
-  }
-
-  /**
-   * Add a single auto-traceroute node.
-   * Keeps branching: MySQL lacks onConflictDoNothing.
-   */
-  async addAutoTracerouteNode(nodeNum: number, sourceId?: string): Promise<void> {
-    const now = this.now();
-    const { autoTracerouteNodes } = this.tables;
-
-    // Pre-check for the (nodeNum, sourceId) tuple. Needed because SQLite/MySQL
-    // treat NULL as distinct in UNIQUE constraints, so insertIgnore would
-    // allow duplicate unscoped rows.
-    const whereClause = sourceId
-      ? and(eq(autoTracerouteNodes.nodeNum, nodeNum), eq(autoTracerouteNodes.sourceId, sourceId))
-      : and(eq(autoTracerouteNodes.nodeNum, nodeNum), isNull(autoTracerouteNodes.sourceId));
-    const existing = await this.db
-      .select({ id: autoTracerouteNodes.id })
-      .from(autoTracerouteNodes)
-      .where(whereClause)
-      .limit(1);
-    if (existing.length > 0) return;
-
-    await this.insertIgnore(autoTracerouteNodes, {
-      nodeNum,
-      createdAt: now,
-      sourceId: sourceId ?? null,
-    });
-  }
-
-  /**
-   * Remove a single auto-traceroute node
-   */
-  async removeAutoTracerouteNode(nodeNum: number, sourceId?: string): Promise<void> {
-    const { autoTracerouteNodes } = this.tables;
-    const where = sourceId
-      ? and(
-          eq(autoTracerouteNodes.nodeNum, nodeNum),
-          eq(autoTracerouteNodes.sourceId, sourceId)
-        )
-      : eq(autoTracerouteNodes.nodeNum, nodeNum);
-    await this.db.delete(autoTracerouteNodes).where(where);
-  }
-
-  // ============ AUTO TIME SYNC NODES ============
-
-  /**
-   * Get all auto time sync nodes.
-   * When sourceId is provided, return only rows scoped to that source.
-   * When omitted, return everything.
-   */
-  async getAutoTimeSyncNodes(sourceId?: string): Promise<number[]> {
-    const { autoTimeSyncNodes } = this.tables;
-    const query = this.db
-      .select({ nodeNum: autoTimeSyncNodes.nodeNum })
-      .from(autoTimeSyncNodes);
-    const results = sourceId
-      ? await query
-          .where(eq(autoTimeSyncNodes.sourceId, sourceId))
-          .orderBy(asc(autoTimeSyncNodes.createdAt))
-      : await query.orderBy(asc(autoTimeSyncNodes.createdAt));
-    return results.map((r: any) => Number(r.nodeNum));
-  }
-
-  /**
-   * Set auto time sync nodes (replaces all existing entries for the given
-   * source, or globally when sourceId is omitted).
-   */
-  async setAutoTimeSyncNodes(nodeNums: number[], sourceId?: string): Promise<void> {
-    const now = this.now();
-    const { autoTimeSyncNodes } = this.tables;
-
-    // Delete existing entries scoped to this source (or all when unscoped).
-    if (sourceId) {
-      await this.db
-        .delete(autoTimeSyncNodes)
-        .where(eq(autoTimeSyncNodes.sourceId, sourceId));
-    } else {
-      await this.db.delete(autoTimeSyncNodes);
-    }
-    // Insert new entries
-    for (const nodeNum of nodeNums) {
-      await this.db
-        .insert(autoTimeSyncNodes)
-        .values({ nodeNum, createdAt: now, sourceId: sourceId ?? null });
-    }
-  }
-
-  /**
-   * Add a single auto time sync node.
-   * Keeps branching: MySQL lacks onConflictDoNothing.
-   */
-  async addAutoTimeSyncNode(nodeNum: number, sourceId?: string): Promise<void> {
-    const now = this.now();
-    const { autoTimeSyncNodes } = this.tables;
-
-    await this.insertIgnore(autoTimeSyncNodes, {
-      nodeNum,
-      createdAt: now,
-      sourceId: sourceId ?? null,
-    });
-  }
-
-  /**
-   * Remove a single auto time sync node
-   */
-  async removeAutoTimeSyncNode(nodeNum: number, sourceId?: string): Promise<void> {
-    const { autoTimeSyncNodes } = this.tables;
-    const where = sourceId
-      ? and(
-          eq(autoTimeSyncNodes.nodeNum, nodeNum),
-          eq(autoTimeSyncNodes.sourceId, sourceId)
-        )
-      : eq(autoTimeSyncNodes.nodeNum, nodeNum);
-    await this.db.delete(autoTimeSyncNodes).where(where);
-  }
-
-  // ============ CUSTOM THEMES ============
-
-  /**
-   * Normalize a raw theme row into DbCustomTheme format.
-   * Ensures is_builtin is coerced to 0/1 for consistency across dialects.
-   */
-  private normalizeThemeRow(row: any): DbCustomTheme {
-    return {
-      id: Number(row.id),
-      name: row.name,
-      slug: row.slug,
-      definition: row.definition,
-      is_builtin: row.is_builtin ? 1 : 0,
-      created_by: row.created_by != null ? Number(row.created_by) : undefined,
-      created_at: Number(row.created_at),
-      updated_at: Number(row.updated_at),
-    };
-  }
-
-  /**
-   * Get all custom themes ordered by name
-   */
-  async getAllCustomThemes(): Promise<DbCustomTheme[]> {
-    const { customThemes } = this.tables;
-    try {
-      const results = await this.db
-        .select()
-        .from(customThemes)
-        .orderBy(asc(customThemes.name));
-      return results.map((row: any) => this.normalizeThemeRow(row));
-    } catch (error) {
-      logger.error('[MiscRepository] Failed to get custom themes:', error);
-      throw error;
-    }
-  }
-
-  /**
-   * Get a specific custom theme by slug
-   */
-  async getCustomThemeBySlug(slug: string): Promise<DbCustomTheme | undefined> {
-    const { customThemes } = this.tables;
-    try {
-      const results = await this.db
-        .select()
-        .from(customThemes)
-        .where(eq(customThemes.slug, slug))
-        .limit(1);
-      if (results.length === 0) return undefined;
-      return this.normalizeThemeRow(results[0]);
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to get custom theme ${slug}:`, error);
-      throw error;
-    }
-  }
-
-  /**
-   * Create a new custom theme
-   */
-  async createCustomTheme(name: string, slug: string, definitionJson: string, userId?: number): Promise<DbCustomTheme> {
-    const now = Math.floor(Date.now() / 1000);
-    const { customThemes } = this.tables;
-
-    try {
-      if (this.isMySQL()) {
-        // MySQL: use raw query for RETURNING-like behavior via insertId
-        const db = this.getMysqlDb();
-        const result = await db
-          .insert(customThemes)
-          .values({
-            name,
-            slug,
-            definition: definitionJson,
-            is_builtin: false,
-            created_by: userId ?? null,
-            created_at: now,
-            updated_at: now,
-          });
-        const id = Number((result as any)[0].insertId);
-        logger.debug(`[MiscRepository] Created custom theme: ${name} (slug: ${slug})`);
-        return { id, name, slug, definition: definitionJson, is_builtin: 0, created_by: userId, created_at: now, updated_at: now };
-      } else if (this.isPostgres()) {
-        // PostgreSQL: use returning()
-        const db = this.getPostgresDb();
-        const result = await db
-          .insert(customThemes)
-          .values({
-            name,
-            slug,
-            definition: definitionJson,
-            is_builtin: false,
-            created_by: userId ?? null,
-            created_at: now,
-            updated_at: now,
-          })
-          .returning({ id: customThemes.id });
-        const id = Number(result[0].id);
-        logger.debug(`[MiscRepository] Created custom theme: ${name} (slug: ${slug})`);
-        return { id, name, slug, definition: definitionJson, is_builtin: 0, created_by: userId, created_at: now, updated_at: now };
-      } else {
-        // SQLite: use returning()
-        const db = this.getSqliteDb();
-        const result = await db
-          .insert(customThemes)
-          .values({
-            name,
-            slug,
-            definition: definitionJson,
-            is_builtin: false,
-            created_by: userId ?? null,
-            created_at: now,
-            updated_at: now,
-          })
-          .returning({ id: customThemes.id });
-        const id = Number(result[0].id);
-        logger.debug(`[MiscRepository] Created custom theme: ${name} (slug: ${slug})`);
-        return { id, name, slug, definition: definitionJson, is_builtin: 0, created_by: userId, created_at: now, updated_at: now };
-      }
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to create custom theme ${name}:`, error);
-      throw error;
-    }
-  }
-
-  /**
-   * Update an existing custom theme
-   */
-  async updateCustomTheme(slug: string, updates: Partial<{ name: string; definition: string }>): Promise<boolean> {
-    const { customThemes } = this.tables;
-
-    try {
-      // Check if theme exists
-      const existing = await this.getCustomThemeBySlug(slug);
-      if (!existing) {
-        logger.warn(`[MiscRepository] Cannot update non-existent theme: ${slug}`);
-        return false;
-      }
-
-      const now = Math.floor(Date.now() / 1000);
-      const setData: Record<string, any> = { updated_at: now };
-
-      if (updates.name !== undefined) {
-        setData.name = updates.name;
-      }
-      if (updates.definition !== undefined) {
-        setData.definition = typeof updates.definition === 'string'
-          ? updates.definition
-          : JSON.stringify(updates.definition);
-      }
-
-      await this.db
-        .update(customThemes)
-        .set(setData)
-        .where(eq(customThemes.slug, slug));
-
-      logger.debug(`[MiscRepository] Updated custom theme: ${slug}`);
-      return true;
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to update custom theme ${slug}:`, error);
-      throw error;
-    }
-  }
-
-  /**
-   * Delete a custom theme by slug
-   */
-  async deleteCustomTheme(slug: string): Promise<boolean> {
-    const { customThemes } = this.tables;
-
-    try {
-      // Check if theme exists and is not built-in
-      const existing = await this.getCustomThemeBySlug(slug);
-      if (!existing) {
-        logger.warn(`[MiscRepository] Cannot delete non-existent theme: ${slug}`);
-        return false;
-      }
-      if (existing.is_builtin) {
-        throw new Error('Cannot delete built-in themes');
-      }
-
-      await this.db
-        .delete(customThemes)
-        .where(eq(customThemes.slug, slug));
-
-      logger.debug(`[MiscRepository] Deleted custom theme: ${slug}`);
-      return true;
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to delete custom theme ${slug}:`, error);
-      throw error;
-    }
-  }
-
-  // ============ DISTANCE DELETE LOG ============
-
-  /**
-   * Get auto-delete-by-distance log entries.
-   * If sourceId is provided, scope to that source. NULL sourceId is the
-   * legacy/unscoped bucket; an unfiltered call returns all rows.
-   */
-  async getDistanceDeleteLog(limit: number = 10, sourceId?: string): Promise<any[]> {
-    const { autoDistanceDeleteLog } = this.tables;
-    const baseQuery = (this.db as any)
-      .select()
-      .from(autoDistanceDeleteLog);
-    const query = sourceId
-      ? baseQuery.where(eq(autoDistanceDeleteLog.sourceId, sourceId))
-      : baseQuery;
-    const rows = await query
-      .orderBy(desc(autoDistanceDeleteLog.timestamp))
-      .limit(limit);
-    return (rows as any[]).map((e: any) => ({
-      ...e,
-      details: e.details ? JSON.parse(e.details) : [],
-    }));
-  }
-
-  /**
-   * Add an entry to the auto-delete-by-distance log
-   */
-  async addDistanceDeleteLogEntry(entry: {
-    timestamp: number;
-    nodesDeleted: number;
-    thresholdKm: number;
-    details: string;
-    sourceId?: string;
-  }): Promise<void> {
-    const { autoDistanceDeleteLog } = this.tables;
-    const now = Date.now();
-    await (this.db as any).insert(autoDistanceDeleteLog).values({
-      timestamp: entry.timestamp,
-      nodesDeleted: entry.nodesDeleted,
-      thresholdKm: entry.thresholdKm,
-      details: entry.details,
-      createdAt: now,
-      sourceId: entry.sourceId ?? null,
-    });
   }
 
   // ============ PACKET LOG ============
@@ -898,273 +495,6 @@ export class MiscRepository extends BaseRepository {
     }));
   }
 
-  // ============ auto_traceroute_log async methods (all backends) ============
-
-  /**
-   * Get the most recent auto-traceroute log rows (all backends).
-   * Boolean success is returned (null preserved).
-   */
-  async getAutoTracerouteLog(limit: number = 10, sourceId?: string): Promise<Array<{
-    id: number;
-    timestamp: number;
-    toNodeNum: number;
-    toNodeName: string | null;
-    success: boolean | null;
-  }>> {
-    try {
-      const { autoTracerouteLog } = this.tables;
-      const query = this.db
-        .select({
-          id: autoTracerouteLog.id,
-          timestamp: autoTracerouteLog.timestamp,
-          toNodeNum: autoTracerouteLog.toNodeNum,
-          toNodeName: autoTracerouteLog.toNodeName,
-          success: autoTracerouteLog.success,
-        })
-        .from(autoTracerouteLog);
-      const scoped = sourceId !== undefined
-        ? query.where(eq(autoTracerouteLog.sourceId, sourceId))
-        : query;
-      const rows = await scoped
-        .orderBy(desc(autoTracerouteLog.timestamp))
-        .limit(limit);
-      return (rows as any[]).map((r: any) => ({
-        id: Number(r.id),
-        timestamp: Number(r.timestamp),
-        toNodeNum: Number(r.toNodeNum),
-        toNodeName: r.toNodeName ?? null,
-        success: r.success === null || r.success === undefined ? null : Boolean(r.success),
-      }));
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to get auto traceroute log: ${error}`);
-      return [];
-    }
-  }
-
-  /**
-   * Insert an auto-traceroute attempt row and prune to the last 100 entries.
-   * Returns the inserted row id. Works for SQLite/Postgres/MySQL.
-   */
-  async logAutoTracerouteAttempt(toNodeNum: number, toNodeName: string | null, sourceId?: string): Promise<number> {
-    try {
-      const { autoTracerouteLog } = this.tables;
-      const now = Date.now();
-      const values: any = {
-        timestamp: now,
-        toNodeNum,
-        toNodeName,
-        success: null,
-        createdAt: now,
-        sourceId: sourceId ?? null,
-      };
-
-      let insertedId = 0;
-      if (this.isPostgres()) {
-        const result = await (this.db.insert(autoTracerouteLog).values(values) as any).returning({ id: autoTracerouteLog.id });
-        insertedId = Number((result as any[])[0]?.id ?? 0);
-      } else if (this.isMySQL()) {
-        const result = await this.db.insert(autoTracerouteLog).values(values);
-        insertedId = Number((result as any)?.[0]?.insertId ?? 0);
-      } else {
-        const result = await this.db.insert(autoTracerouteLog).values(values);
-        insertedId = Number((result as any)?.lastInsertRowid ?? 0);
-      }
-
-      // Prune older rows beyond the 100 most recent.
-      const keepRows = await this.db
-        .select({ id: autoTracerouteLog.id })
-        .from(autoTracerouteLog)
-        .orderBy(desc(autoTracerouteLog.timestamp))
-        .limit(100);
-      const keepIds = (keepRows as any[]).map((r) => Number(r.id));
-      if (keepIds.length > 0) {
-        await this.db
-          .delete(autoTracerouteLog)
-          .where(notInArray(autoTracerouteLog.id, keepIds));
-      }
-
-      return insertedId;
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to log auto traceroute attempt: ${error}`);
-      return 0;
-    }
-  }
-
-  /**
-   * Update the most recent pending auto-traceroute log row for a given
-   * destination node across all backends.
-   */
-  async updateAutoTracerouteResultByNode(toNodeNum: number, success: boolean): Promise<void> {
-    try {
-      const { autoTracerouteLog } = this.tables;
-      const rows = await this.db
-        .select({ id: autoTracerouteLog.id })
-        .from(autoTracerouteLog)
-        .where(and(eq(autoTracerouteLog.toNodeNum, toNodeNum), isNull(autoTracerouteLog.success))!)
-        .orderBy(desc(autoTracerouteLog.timestamp))
-        .limit(1);
-      if ((rows as any[]).length > 0) {
-        const id = Number(((rows as any[])[0]).id);
-        await this.db
-          .update(autoTracerouteLog)
-          .set({ success: success ? 1 : 0 } as any)
-          .where(eq(autoTracerouteLog.id, id));
-      }
-    } catch (error) {
-      logger.error(`[MiscRepository] Failed to update auto traceroute result: ${error}`);
-    }
-  }
-
-  // ============ auto_traceroute_nodes sync methods (SQLite only) ============
-
-  /**
-   * Synchronously get the list of auto-traceroute node nums ordered by
-   * creation time ascending (SQLite only).
-   */
-  getAutoTracerouteNodesSync(): number[] {
-    const db = this.getSqliteDb();
-    const { autoTracerouteNodes } = this.tables;
-    const rows = db
-      .select({ nodeNum: autoTracerouteNodes.nodeNum })
-      .from(autoTracerouteNodes)
-      .orderBy(asc(autoTracerouteNodes.createdAt))
-      .all();
-    return (rows as any[]).map((r) => Number(r.nodeNum));
-  }
-
-  /**
-   * Synchronously replace the auto-traceroute nodes set in a single
-   * transaction (SQLite only). Bad nodeNums are skipped.
-   */
-  setAutoTracerouteNodesSync(nodeNums: number[]): void {
-    const db = this.getSqliteDb();
-    const { autoTracerouteNodes } = this.tables;
-    const now = Date.now();
-    db.transaction((tx) => {
-      tx.delete(autoTracerouteNodes).run();
-      for (const nodeNum of nodeNums) {
-        try {
-          tx.insert(autoTracerouteNodes)
-            .values({ nodeNum, createdAt: now } as any)
-            .run();
-        } catch (error) {
-          logger.debug(`Skipping invalid nodeNum: ${nodeNum}`, error);
-        }
-      }
-    });
-  }
-
-  // ============ auto_traceroute_log sync methods (SQLite only) ============
-
-  /**
-   * Synchronously insert an auto-traceroute attempt row and prune to the last
-   * 100 entries (SQLite only). Returns the inserted row id.
-   */
-  logAutoTracerouteAttemptSync(toNodeNum: number, toNodeName: string | null, sourceId?: string): number {
-    const db = this.getSqliteDb();
-    const { autoTracerouteLog } = this.tables;
-    const now = Date.now();
-    const result = db
-      .insert(autoTracerouteLog)
-      .values({
-        timestamp: now,
-        toNodeNum,
-        toNodeName,
-        success: null,
-        createdAt: now,
-        sourceId: sourceId ?? null,
-      } as any)
-      .run() as any;
-
-    // Prune older rows beyond the 100 most recent.
-    const keepRows = db
-      .select({ id: autoTracerouteLog.id })
-      .from(autoTracerouteLog)
-      .orderBy(desc(autoTracerouteLog.timestamp))
-      .limit(100)
-      .all();
-    const keepIds = (keepRows as any[]).map((r) => Number(r.id));
-    if (keepIds.length > 0) {
-      db.delete(autoTracerouteLog)
-        .where(sql`id NOT IN (${sql.join(keepIds.map((id) => sql`${id}`), sql`, `)})`)
-        .run();
-    }
-
-    return Number(result?.lastInsertRowid ?? 0);
-  }
-
-  /**
-   * Synchronously mark an auto-traceroute log row's success flag (SQLite only).
-   */
-  updateAutoTracerouteResultSync(logId: number, success: boolean): void {
-    const db = this.getSqliteDb();
-    const { autoTracerouteLog } = this.tables;
-    db.update(autoTracerouteLog)
-      .set({ success: success ? 1 : 0 } as any)
-      .where(eq(autoTracerouteLog.id, logId))
-      .run();
-  }
-
-  /**
-   * Synchronously update the most recent pending auto-traceroute log row for
-   * a given destination node (SQLite only).
-   */
-  updateAutoTracerouteResultByNodeSync(toNodeNum: number, success: boolean): void {
-    const db = this.getSqliteDb();
-    const { autoTracerouteLog } = this.tables;
-    const rows = db
-      .select({ id: autoTracerouteLog.id })
-      .from(autoTracerouteLog)
-      .where(and(eq(autoTracerouteLog.toNodeNum, toNodeNum), isNull(autoTracerouteLog.success))!)
-      .orderBy(desc(autoTracerouteLog.timestamp))
-      .limit(1)
-      .all();
-    if (rows.length > 0) {
-      const id = Number((rows[0] as any).id);
-      db.update(autoTracerouteLog)
-        .set({ success: success ? 1 : 0 } as any)
-        .where(eq(autoTracerouteLog.id, id))
-        .run();
-    }
-  }
-
-  /**
-   * Synchronously fetch the most recent auto-traceroute log rows (SQLite only).
-   * Boolean success is returned (null preserved).
-   */
-  getAutoTracerouteLogSync(limit: number = 10, sourceId?: string): Array<{
-    id: number;
-    timestamp: number;
-    toNodeNum: number;
-    toNodeName: string | null;
-    success: boolean | null;
-  }> {
-    const db = this.getSqliteDb();
-    const { autoTracerouteLog } = this.tables;
-    const query = db
-      .select({
-        id: autoTracerouteLog.id,
-        timestamp: autoTracerouteLog.timestamp,
-        toNodeNum: autoTracerouteLog.toNodeNum,
-        toNodeName: autoTracerouteLog.toNodeName,
-        success: autoTracerouteLog.success,
-      })
-      .from(autoTracerouteLog);
-    const rows = (sourceId !== undefined
-      ? query.where(eq(autoTracerouteLog.sourceId, sourceId))
-      : query)
-      .orderBy(desc(autoTracerouteLog.timestamp))
-      .limit(limit)
-      .all();
-    return (rows as any[]).map((r: any) => ({
-      id: Number(r.id),
-      timestamp: Number(r.timestamp),
-      toNodeNum: Number(r.toNodeNum),
-      toNodeName: r.toNodeName ?? null,
-      success: r.success === null || r.success === undefined ? null : r.success === 1,
-    }));
-  }
-
   /**
    * Get packet counts per from_node since a given timestamp, excluding internal
    * traffic (packets where both ends are the local node). Used for spam
@@ -1333,128 +663,6 @@ export class MiscRepository extends BaseRepository {
     } catch (error) {
       logger.error('[MiscRepository] Failed to get packet counts by portnum:', error);
       return [];
-    }
-  }
-
-  // ============ USER MAP PREFERENCES ============
-
-  /**
-   * Get map preferences for a user
-   */
-  async getMapPreferences(userId: number): Promise<Record<string, any> | null> {
-    try {
-      const table = this.tables.userMapPreferences;
-      const rows = await this.db
-        .select()
-        .from(table)
-        .where(eq(table.userId, userId))
-        .limit(1);
-
-      if (rows.length === 0) return null;
-      const row = rows[0] as any;
-
-      return {
-        mapTileset: row.mapTileset ?? null,
-        showPaths: row.showPaths ?? false,
-        showNeighborInfo: row.showNeighborInfo ?? false,
-        showRoute: row.showRoute ?? true,
-        showMotion: row.showMotion ?? true,
-        showMqttNodes: row.showMqttNodes ?? true,
-        showUdpNodes: row.showUdpNodes ?? false,
-        showRfNodes: row.showRfNodes ?? true,
-        showMeshCoreNodes: row.showMeshcoreNodes ?? true,
-        showWaypoints: row.showWaypoints ?? true,
-        showAnimations: row.showAnimations ?? false,
-        showAccuracyRegions: row.showAccuracyRegions ?? false,
-        showEstimatedPositions: row.showEstimatedPositions ?? false,
-        positionHistoryHours: row.positionHistoryHours ?? null,
-        mapMaxAgeHours: row.mapMaxAgeHours ?? null,
-        positionHistoryPointsOnly: row.positionHistoryPointsOnly ?? false,
-      };
-    } catch (error) {
-      logger.error('[MiscRepository] Failed to get map preferences:', error);
-      return null;
-    }
-  }
-
-  /**
-   * Save map preferences for a user (upsert).
-   */
-  async saveMapPreferences(userId: number, preferences: {
-    mapTileset?: string;
-    showPaths?: boolean;
-    showNeighborInfo?: boolean;
-    showRoute?: boolean;
-    showMotion?: boolean;
-    showMqttNodes?: boolean;
-    showUdpNodes?: boolean;
-    showRfNodes?: boolean;
-    showMeshCoreNodes?: boolean;
-    showWaypoints?: boolean;
-    showAnimations?: boolean;
-    showAccuracyRegions?: boolean;
-    showEstimatedPositions?: boolean;
-    positionHistoryHours?: number | null;
-    mapMaxAgeHours?: number | null;
-    positionHistoryPointsOnly?: boolean;
-  }): Promise<void> {
-    try {
-      const table = this.tables.userMapPreferences;
-      const existing = await this.db
-        .select({ id: table.id })
-        .from(table)
-        .where(eq(table.userId, userId))
-        .limit(1);
-
-      if (existing.length > 0) {
-        const set: Record<string, any> = {};
-        if (preferences.mapTileset !== undefined) set.mapTileset = preferences.mapTileset;
-        if (preferences.showPaths !== undefined) set.showPaths = preferences.showPaths;
-        if (preferences.showNeighborInfo !== undefined) set.showNeighborInfo = preferences.showNeighborInfo;
-        if (preferences.showRoute !== undefined) set.showRoute = preferences.showRoute;
-        if (preferences.showMotion !== undefined) set.showMotion = preferences.showMotion;
-        if (preferences.showMqttNodes !== undefined) set.showMqttNodes = preferences.showMqttNodes;
-        if (preferences.showUdpNodes !== undefined) set.showUdpNodes = preferences.showUdpNodes;
-        if (preferences.showRfNodes !== undefined) set.showRfNodes = preferences.showRfNodes;
-        if (preferences.showMeshCoreNodes !== undefined) set.showMeshcoreNodes = preferences.showMeshCoreNodes;
-        if (preferences.showWaypoints !== undefined) set.showWaypoints = preferences.showWaypoints;
-        if (preferences.showAnimations !== undefined) set.showAnimations = preferences.showAnimations;
-        if (preferences.showAccuracyRegions !== undefined) set.showAccuracyRegions = preferences.showAccuracyRegions;
-        if (preferences.showEstimatedPositions !== undefined) set.showEstimatedPositions = preferences.showEstimatedPositions;
-        if (preferences.positionHistoryHours !== undefined) set.positionHistoryHours = preferences.positionHistoryHours;
-        if (preferences.mapMaxAgeHours !== undefined) set.mapMaxAgeHours = preferences.mapMaxAgeHours;
-        if (preferences.positionHistoryPointsOnly !== undefined) set.positionHistoryPointsOnly = preferences.positionHistoryPointsOnly;
-
-        if (Object.keys(set).length > 0) {
-          await this.db.update(table).set(set).where(eq(table.userId, userId));
-        }
-      } else {
-        const now = Date.now();
-        await this.db.insert(table).values({
-          userId,
-          mapTileset: preferences.mapTileset ?? null,
-          showPaths: preferences.showPaths ?? false,
-          showNeighborInfo: preferences.showNeighborInfo ?? false,
-          showRoute: preferences.showRoute ?? true,
-          showMotion: preferences.showMotion ?? true,
-          showMqttNodes: preferences.showMqttNodes ?? false,
-          showUdpNodes: preferences.showUdpNodes ?? false,
-          showRfNodes: preferences.showRfNodes ?? true,
-          showMeshcoreNodes: preferences.showMeshCoreNodes ?? true,
-          showWaypoints: preferences.showWaypoints ?? true,
-          showAnimations: preferences.showAnimations ?? false,
-          showAccuracyRegions: preferences.showAccuracyRegions ?? false,
-          showEstimatedPositions: preferences.showEstimatedPositions ?? true,
-          positionHistoryHours: preferences.positionHistoryHours ?? null,
-          mapMaxAgeHours: preferences.mapMaxAgeHours ?? null,
-          positionHistoryPointsOnly: preferences.positionHistoryPointsOnly ?? false,
-          createdAt: now,
-          updatedAt: now,
-        });
-      }
-    } catch (error) {
-      logger.error('[MiscRepository] Failed to save map preferences:', error);
-      throw error;
     }
   }
